tweet-sentiment-hadoop/requirements.txt
# Core
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0

# NLP & ML
scikit-learn>=1.1.0
nltk>=3.7

# Spark (PySpark)
pyspark>=3.3.0

# Kafka
kafka-python>=2.0.2

# Visualization
streamlit>=1.25.0
plotly>=5.14.0
seaborn>=0.12.0
matplotlib>=3.6.0

# Optional: Deep Learning
# tensorflow>=2.11.0
# torch>=2.0.0
